Remove deprecated test
