+++ tests/requirements.txt
@@ -1,2 +1,1 @@
-requests==2.30.0
 validators==0.20.0
